[DEV APPROVED] Add danger to enforce Mas Standards & Code Quality

+++ Jenkinsfile
@@ -11,10 +11,22 @@ pipeline {
             sh "docker-compose -f docker-compose.yml up -d"
           }
         }
-        stage('test') {
-          steps {
-            sh "docker-compose -f docker-compose.yml run --rm rails ./script/test"
-          }
+        stage ('branch-test') {
+          when { not { branch 'PR-*' } }
+            steps {
+                sh "docker-compose -f docker-compose.yml run --rm rails ./script/test"
+            }
+        }
+        stage ('pr-test') {
+          when { branch 'PR-*' }
+            environment {
+              DANGER_CHANGE_ID = "${env.CHANGE_ID}"
+              DANGER_CHANGE_URL = "${env.CHANGE_URL}"
+              DANGER_JENKINS_URL = "${env.JENKINS_URL}"
+            }
+            steps {
+                sh "docker-compose -f docker-compose.yml run --rm rails ./script/test"
+            }
         }
     }
     post {
